Add Git hook to prepend JIRA ticket number to commit messages
This commit adds a new install script that builds and installs a Go script for adding JIRA ticket numbers to commit messages. The Go script is integrated with Lefthook, which checks if the commit message contains a JIRA ticket number. If not, it prepends the ticket number derived from the branch name to the beginning of the message.

The README file has also been updated with instructions on how to set up and use this feature.